+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add the `empty` event to FunctionQueue
 * Add the `RURL#extension` property to get/set the extension of the pathname
 * `RURL#pathname` will now no longer be empty when the protocol requires a slash
+* Add a custom `Error` class
 
 ## 0.7.0 (2020-03-13)